Adding support for BatchGetItems/Scan/Query
Created a ReadIsolationHandler interface that takes an item
and handles it according to the isolation level of the handler.
There are two implementations of ReadIsolationHandlers:
* ReadCommittedIsolationHandlerImpl will only return a
committed item (possibly an older version stored in the items table).
The logic for this class was refactored from
Transaction.getItemCommitted(...).
* ReadUncommittedIsolationHandlerImpl only removes transient items.

The ReadIsolationHandlers are used to support batchGetItems, scan, and
query in the TransactionManagerDynamoDBFacade at the UNCOMMITTED and
COMMITTED isolation levels.